python-fastapi/pyproject.toml
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[project]
requires-python = ">=3.10"
name = "speed-knowledge-server"
version = "0.1.0"
description = "Speed Knowledge Server"
authors = [
    { name = "ykx" }
]
dependencies = [
  "fastapi==0.109.0",
  "uvicorn[standard]==0.27.0",
  "pydantic>=2.12.0",
  "pydantic-settings>=2.7.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.13.1",
  "httpx==0.27.2",
  "pymysql==1.1.0",
  "python-jose[cryptography]==3.3.0",
  "passlib==1.7.4",
  "bcrypt>=3.2.0,<4.0.0",
  "python-multipart==0.0.6",
  "jinja2==3.1.2",
  "email-validator==2.1.0",
  "python-dotenv==1.0.0",
  "captcha==0.5.0",
  "redis==5.0.1",
  "minio==7.2.18",
  "apscheduler==3.10.4",
  "volcengine-python-sdk==5.0.6",
  "langchain>=1.2.12",
  "langchain-text-splitters>=1.1.1",
  "langchain-qdrant>=1.1.0",
  "qdrant-client>=1.17.1",
  "langchain-huggingface>=1.2.1",
  "dashscope>=1.25.14",
  "aio-pika>=9.6.1",
  "langchain-community>=0.4.1",
  "sentence-transformers>=5.3.0",
  "langsmith>=0.7.18",
]
[dependency-groups]
dev = [
  "pytest==7.4.4",
  "pytest-asyncio==0.23.3",
  "pytest-cov==4.1.0",

  "black==23.12.1",
  "isort==5.13.2",
  "flake8==7.0.0",
  "mypy==1.8.0",
  "pylint==3.0.3",

  "types-passlib==1.7.7.20240106"
]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

# 忽略第三方库的类型检查
[[tool.mypy.overrides]]
module = ["redis.*", "fastapi.*", "sqlalchemy.*", "captcha.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

